Do the filtering of result of `hotdog ssh` using `filter_hosts()` in favor of `rewrite_expression()`
Rewriting `options[:index]` as SLICE expression won't work as expected with hosts' status
since the result may be filtered again with using the status, the filtering needs to be
done after the `evaluate()`.

For now we need to keep using `filter_hosts()` in favor of `rewrite_expression() to do
the filtering based on status filtering.